Added in by hand sum function in `BasicMathService`
Added in a by hand sum function in the `BasicMathService`. This takes two strings, goes through them adding from the far right of the string until the string with less characters is done. Then passes the remainder and the rest of the larger string to the same function to keep adding until everything has been accounted for.

Added in tests for this similar to the other tests. Didn't do the localization like I did with the BigDecimal solution in `MathService`.